[BUG-120481][SPARK-24785][SHELL] Making sure REPL prints Spark UI info and then Welcome message
After https://github.com/apache/spark/pull/21495 the welcome message is printed first, and then Scala prompt will be shown before the Spark UI info is printed.

Although it's a minor issue, but visually, it doesn't look as nice as the existing behavior. This PR intends to fix it by duplicating the Scala `process` code to arrange the printing order. However, one variable is private, so reflection has to be used which is not desirable.

We can use this PR to brainstorm how to handle it properly and how Scala can change their APIs to fit our need.

Existing test

Closes #21749 from dbtsai/repl-followup.

Change-Id: I119fc51b729f684cc42e11978489757cbf899998
Authored-by: DB Tsai <d_tsai@apple.com>
Signed-off-by: DB Tsai <d_tsai@apple.com>