[SPARK-25482][SQL] Avoid pushdown of subqueries to data source filters
An expressions with a subquery can be pushed down as a data source filter. Despite the filter is not actively used, this causes anyway a re-execution of the subquery becuase the `ReuseSubquery` optimization rule is ineffective in this case.

The PR avoids this problem by forbidding the push down of filters containing a subquery.

added UT

Closes #22518 from mgaido91/SPARK-25482.

Change-Id: I93e50b368c25fbf1b98495d5c20aac4785a55aec
Authored-by: Marco Gaido <marcogaido91@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>