[CDPD-22118][SPARK-26700][CORE] enable fetch-big-block-to-disk by default
This is a followup of #16989

The fetch-big-block-to-disk feature is disabled by default, because it's not compatible with external shuffle service prior to Spark 2.2. The client sends stream request to fetch block chunks, and old shuffle service can't support it.

After 2 years, Spark 2.2 has EOL, and now it's safe to turn on this feature by default

existing tests

Closes #23625 from cloud-fan/minor.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>
(cherry picked from commit ed71a825c56920327533ebb741707871848ccd6d)
Change-Id: I0d6708d922136a5aa88f78586d5bc9096a8b8826