CDPD-414. Lightweight dynamic allocation without a shuffle service.
This is an experimental feature to provide dynamic allocation without
the need for external shuffle storage.

It follows the same precedent set by special-casing executors that
contain cached blocks, but also takes into account shuffle data. So
if an executor stores shuffle data, there's a separate idle timeout
for that executor.

There are a couple of changes to how executors are timed out that make
this more efficient than the current cache timeout; mainly, if the
shuffle (or cache) data is cleaned up by Spark, the executor now
properly accounts for the idle timeout, instead of staying alive
for the longer timeout period like before. This would be something to
also fix in the main path, but in the interest of not changing
existing behavior, is kept under the new flag.

Ultimately, this is a balancing act: it doesn't need infrastructure
for shuffle storage outside of executors, at the expense of keeping
executors around for longer than before. So to achieve a reasonable
level of elasticity, users would have to play with the timeouts and
with things like spark.cleaner.periodicGC.interval, which by default
is 30min. That is a little bit harder on the python side, since there
isn't a configuration to force python's gc to run periodically, which
may cause shuffle (and cache) blocks to remain alive for longer than
necessary.

(cherry picked from commit 8027781faf44554d2be33c4c77eb84f031e73950)

Change-Id: I7e9f6adc47b4b73903a5f63df4fda948a04d87d6