[SPARK-26089][CORE] Handle corruption in large shuffle blocks
SPARK-4105 added corruption detection in shuffle blocks but that was limited to
blocks which are smaller than maxBytesInFlight/3. This commit adds upon that by
adding corruption check for large blocks. There are two changes/improvements
that are made in this commit:

1. Large blocks are checked upto maxBytesInFlight/3 size in a similar way as
smaller blocks, so if a large block is corrupt in the starting, that block will
be re-fetched and if that also fails, FetchFailureException will be thrown.
2. If large blocks are corrupt after size maxBytesInFlight/3, then any
IOException thrown while reading the stream will be converted to
FetchFailureException.  This is slightly more aggressive than was originally
intended but since the consumer of the stream may have already read some records
and processed them, we can't just re-fetch the block, we need to fail the whole
task. Additionally, we also thought about maybe adding a new type of
TaskEndReason, which would re-try the task couple of times before failing the
previous stage, but given the complexity involved in that solution we decided to
not proceed in that direction.

Thanks to squito for direction and support.

Changed the junit test for big blocks to check for corruption.

Closes #23453 from ankuriitg/ankurgupta/SPARK-26089.

Cloudera ID: CDH-75631

Change-Id: I554598161a7ac5dd09c91f050de909646206b12f
Authored-by: ankurgupta <ankur.gupta@cloudera.com>
Signed-off-by: Imran Rashid <irashid@cloudera.com>
(cherry picked from commit cab6371fcac69f9507299522937fcbe5b334c407)