CDH-80437. Recognize S3 URIs correctly, don't emit lineage for unknown URIs.
The first issue is that S3 URIs were not correctly identified; the fix is
simple, most of the rest is mocking S3 for testing.

The second issue is to not emit any lineage when the URIs are not recognized,
because Navigator cannot deal with those and will just leave the lineage
files around instead.

(cherry picked from commit 7c28ca31f8fc303d6832c20398c2a1eb958310dd)
Change-Id: I680fa897ea61f4c31c511d5f1356252de3f5f3c7