[SPARK-26895][CORE][2.4] prepareSubmitEnvironment should be called within doAs for proxy users
This also includes the follow-up fix by Hyukjin Kwon:
[SPARK-26895][CORE][FOLLOW-UP] Uninitializing log after `prepareSubmitEnvironment` in SparkSubmit

`prepareSubmitEnvironment` performs globbing that will fail in the case where a proxy user (`--proxy-user`) doesn't have permission to the file. This is a bug also with 2.3, so we should backport, as currently you can't launch an application that for instance is passing a file under `--archives`, and that file is owned by the target user.

The solution is to call `prepareSubmitEnvironment` within a doAs context if proxying.

Manual tests running with `--proxy-user` and `--archives`, before and after, showing that the globbing is successful when the resource is owned by the target user.

I've looked at writing unit tests, but I am not sure I can do that cleanly (perhaps with a custom FileSystem). Open to ideas.

Closes #25516 from vanzin/SPARK-26895-2.4.

Authored-by: Alessandro Bellina <abellina@gmail.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>
(cherry picked from commit fd2fe15b393660974777251034e4a8a3e320b5d0)

Cloudera ID: CDH-82996

Change-Id: I0ff6e18dc587754b4fa92d83dff8bda5b14787f0