[SPARK-28836][SQL] Remove the canonicalize(attributes) method from PlanExpression
This PR removes the `canonicalize(attrs: AttributeSeq)` from `PlanExpression` and taking care of normalizing expressions in `QueryPlan`.

`Expression` has already a `canonicalized` method and having the `canonicalize` method in `PlanExpression` is confusing.

Removes the `canonicalize` plan from `PlanExpression`. Also renames the `normalizeExprId` to `normalizeExpressions` in query plan.

This PR is a refactoring and passes the existing tests

Closes #25534 from dbaliafroozeh/ImproveCanonicalizeAPI.

Change-Id: I14e6a6dbd14d34926a36ffda3de8af2022dc8593
Authored-by: Ali Afroozeh <ali.afroozeh@databricks.com>
Signed-off-by: herman <herman@databricks.com>