CDPD-3881. Add required changes for upgrading the built-in hive to cdp hive
This commit contains several changes in order to upgrade the built-in hive:
Make changes in a few places to conisder the new built-in  version.
Set the hive execution engine to mr to avoid hive from checking for tez configs
Bring back the calcite dependency, since it is needed for cdp hive after built-in hive upgarde
Remove the usages of hive23.versions since in CDP we just care about one hive version
Make changes to different tests to make them work with the new hive version or ignore them. The most notable one was to create a replacement test in HiveUDFSuite that doesn't use timestamp.

Ignore the tests that creates indexes, since the support for index is removed in cdp hive.

Ignore some failing unit tests in sql/hive-thrifserver module, sine we don't want to support triftserver in cdp.

Add internal repo to download hive-contrib and hive-hctatlog-core jars for testing

Ignore tests because of CDPD-7882 and CDPD-8431.

Update the hive version to the recent cdp hive cdpd-master branch

Also add required changes to sql/hive-thriftserver module to upgrade built in hive to cdp hive

Thanks to Imran Rashid and Bruce Robbins who helped with testing, providing fixes for test failures, and also reviewing the change.

Change-Id: I88c43887bc3a11fcb182a3fb8371db484b511936