[SPARK-30129][CORE][2.4] Set client's id in TransportClient after successful auth
The new auth code was missing this bit, so it was not possible to know which
app a client belonged to when auth was on.

I also refactored the SASL test that checks for this so it also checks the
new protocol (test failed before the fix, passes now).

Closes #26764 from vanzin/SPARK-30129-2.4.

Authored-by: Marcelo Vanzin <vanzin@cloudera.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>
(cherry picked from commit 3b64e2f35657ef0a4001a4f0926ead8cd9226a28)

Cloudera ID: CDPD-6419

Change-Id: I3ef56d96c1ae156f2169c7d3ec847291f0d766ba