[SPARK-30240][core] Support HTTP redirects directly to a proxy server.
Spark's web UI has support for generating links to paths with a
prefix, to support a proxy server, but those do not apply when
the UI is responding with redirects. In that case, Spark is sending
its own URL back to the client, and if it's behind a dumb proxy
server that doesn't do rewriting (like when using stunnel for HTTPS
support) then the client will see the wrong URL and may fail.

The change adds a new config option to configure an address for the
proxy server, and a new handler that intercepts redirects and replaces
the URL with one pointing at the proxy server. This is needed on top
of the "proxy base path" support because redirects use full URLs, not
just absolute paths from the server's root.

Tested with added unit test, with Spark behind stunnel, and in a
more complicated app using a different HTTPS proxy.

Cloudera ID: CDPD-6540

Change-Id: I0540b2fcaaea8109f913e2c0c8a113e829e12d7d